+++ src/GroupProject/Models/Konto.cs
@@ -13,8 +13,8 @@ namespace GroupProject.Models
         [Required]
         public string kontoNr { get; set; }
         [Required]
-        [Range(-10000, 10000000)]
-        public int saldo { get; set; }
+        [Range(-10000.00, 10000000.00)]
+        public decimal saldo { get; set; }
 
         public virtual Person Personer { get; set; }
         [Required]
